+++ platform/Codat/Platform/Models/Components/WebhookConsumerPrototype.cs
@@ -12,6 +12,7 @@ namespace Codat.Platform.Models.Components
     using Codat.Platform.Utils;
     using Newtonsoft.Json;
     using System.Collections.Generic;
+    using System;
     
     public class WebhookConsumerPrototype
     {
@@ -35,8 +36,15 @@ namespace Codat.Platform.Models.Components
         public List<string>? EventTypes { get; set; }
 
         /// <summary>
+        /// Company tags provide an additional way to filter messages, independent of event types. Company tags are case-sensitive, and only messages from companies with matching tags will be sent to this endpoint. Use the format `tagKey:tagValue`.
+        /// </summary>
+        [JsonProperty("companyTags")]
+        public List<string>? CompanyTags { get; set; }
+
+        /// <summary>
         /// Unique identifier of the company to indicate company-specific events. The associated webhook consumer will receive events only for the specified ID.
         /// </summary>
+        [Obsolete("This field will be removed in a future release, please migrate away from it as soon as possible")]
         [JsonProperty("companyId")]
         public string? CompanyId { get; set; } = null;
     }